Functional Enhancements and Version Updates in WordPress Plugins Changes made in this series of commits:
1. **Inclusion of Additional Function Files for WordPress Plugin**:
   - Added essential function files to the WordPress plugin, expanding its capabilities and features.

2. **Plugin Version Update to 1.2.3 in SecureWeb Headers Optimizer**:
   - Updated the version number from 1.2.2 to 1.2.3 in secureweb-headers-optimizer.php, marking a new release.
   - Removed the 'Platform' header from the plugin file to enhance security by preventing potential information disclosure about the server platform.

3. **Update Theme Version to 1.0.2**:
   - Incremented the version number to 1.0.2 in the style.css file, reflecting the latest updates and improvements in the theme.

4. **Enable SVG Uploads for Administrator Users**:
   - Modified the upload_mimes and wp_check_filetype_and_ext filters to allow SVG file uploads for administrator users, enhancing the flexibility and content management capabilities of the plugin.

5. **Addition of disable-comments.php File in Payvalida Theme**:
   - Introduced a new file, disable-comments.php, to the payvalida-theme/includes directory.
   - This file adds functionality to disable comments on the front-end, hide existing comments, and remove comments-related options from the admin menu and admin bar.

Purpose of These Changes:
These updates collectively aim to enhance the functionality and security of various WordPress plugins and themes. The addition of function files and the version updates signify improvements and new features in the plugins. The removal of the 'Platform' header and the addition of the disable-comments.php file contribute to a more secure and streamlined user experience. Enabling SVG uploads for administrators adds flexibility in content management without compromising security.

Security Considerations:
No security risks are associated with these changes. The updates focus on enhancing functionality and security. The removal of the 'Platform' header reduces potential information disclosure risks. The SVG upload feature is restricted to administrator users to maintain a secure environment. The disable-comments.php file adds an additional layer of security by limiting unwanted interactions on the website.